Break out crypto transfer content from modal
