+++ packages/web/src/components/usdc-manual-transfer-modal/USDCManualTransferModal.module.css
@@ -11,44 +11,3 @@
   padding: 0;
   border: none;
 }
-
-.content {
-  display: flex;
-  flex-direction: column;
-  gap: var(--unit-6);
-}
-
-.buttonContainer {
-  display: flex;
-  flex-direction: row;
-  gap: var(--unit-2);
-}
-
-.buttonContainer.mobile {
-  flex-direction: column;
-}
-
-.data {
-  display: flex;
-  flex-direction: column;
-  gap: var(--unit-6);
-}
-
-.qr {
-  width: 160px;
-  height: 160px;
-  padding: 10px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.columns {
-  display: flex;
-  gap: var(--unit-4);
-  align-items: center;
-}
-
-.columns.mobile {
-  flex-direction: column;
-}
