+++ packages/web/src/components/usdc-manual-transfer/USDCManualTransfer.module.css
@@ -1,0 +1,40 @@
+.root {
+  display: flex;
+  flex-direction: column;
+  gap: var(--unit-6);
+}
+
+.buttonContainer {
+  display: flex;
+  flex-direction: row;
+  gap: var(--unit-2);
+}
+
+.buttonContainer.mobile {
+  flex-direction: column;
+}
+
+.data {
+  display: flex;
+  flex-direction: column;
+  gap: var(--unit-6);
+}
+
+.qr {
+  width: 160px;
+  height: 160px;
+  padding: 10px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.columns {
+  display: flex;
+  gap: var(--unit-4);
+  align-items: center;
+}
+
+.columns.mobile {
+  flex-direction: column;
+}
